+++ src/Twilio/Rest/Api/V2010/Account/Conference/ParticipantList.php
@@ -86,6 +86,7 @@ class ParticipantList extends ListResource {
             'Byoc' => $options['byoc'],
             'CallerId' => $options['callerId'],
             'CallReason' => $options['callReason'],
+            'RecordingTrack' => $options['recordingTrack'],
         ]);
 
         $payload = $this->version->create('POST', $this->uri, [], $data);