+++ src/Twilio/Rest/Video/V1/RecordingSettingsOptions.php
@@ -74,7 +74,7 @@ class CreateRecordingSettingsOptions extends Options {
     }
 
     /**
-     * The URL of the AWS S3 bucket where the recordings should be stored. We only support DNS-compliant URLs like `https://<my-bucket>.s3-<aws-region>.amazonaws.com/recordings`, where `recordings` is the path in which you want the recordings to be stored. This URL accepts only URI-valid characters, as described in the <a href='https://tools.ietf.org/html/rfc3986#section-2'>RFC 3986</a>.
+     * The URL of the AWS S3 bucket where the recordings should be stored. We only support DNS-compliant URLs like `https://documentation-example-twilio-bucket/recordings`, where `recordings` is the path in which you want the recordings to be stored. This URL accepts only URI-valid characters, as described in the <a href='https://tools.ietf.org/html/rfc3986#section-2'>RFC 3986</a>.
      *
      * @param string $awsS3Url The URL of the AWS S3 bucket where the recordings
      *                         should be stored