+++ CHANGES.md
@@ -1,6 +1,16 @@
 twilio-php Changelog
 ====================
 
+[2020-12-08] Version 6.15.0
+---------------------------
+**Api**
+- Added optional `RecordingTrack` parameter for create calls, create participants, and create call recordings
+- Removed deprecated Programmable Chat usage record categories **(breaking change)**
+
+**Twiml**
+- Add `recordingTrack` to `Dial`.
+
+
 [2020-12-02] Version 6.14.0
 ---------------------------
 **Api**